requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.1
PyPDF2==3.0.1
Pillow==10.1.0
itsdangerous==2.1.2
Werkzeug==3.0.1